+++ packages/common/src/relay/calling/results/DisconnectResult.ts
@@ -1,0 +1,8 @@
+import { BaseResult } from './BaseResult'
+import { Disconnect } from '../components/Disconnect'
+
+export class DisconnectResult extends BaseResult {
+  constructor(public component: Disconnect) {
+    super(component)
+  }
+}