+++ packages/node/CHANGELOG.md
@@ -3,7 +3,10 @@ All notable changes to this project will be documented in this file.
 
 This project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
-<!-- ## [Unreleased] -->
+## [Unreleased]
+### Added
+- Call `disconnect()` method.
+
 ## [2.3.1] - 2019-11-01
 ### Fixed
 - Relay client reconnection issue after network failure/reset.